begin refactor (migrate last visit)
