+++ src/context/UserDataContext/old-properties/hideDifficulty.ts
